==> 11-find_closest.ipynb <==
cur_path = '/Users/Joey/Projects/NBA_player_tracking/data/'

import pandas as pd
import numpy as np

shot_attempts_all = pd.read_csv(cur_path+'shot_attempts_all_dist.csv')
shot_attempts_all.head()

team_id_list = list(shot_attempts_all['shot_team_id'])
closest_player = []
closest_dist = []
closest_x = []
closest_y = []
for i in range(len(shot_attempts_all)):
  shooter_team = team_id_list[i]
  min_dist = 9999
  for j in range(-10,0,1):
    if j == -1:
      p_id = shot_attempts_all.columns[j][0:3]
    else:
      p_id = shot_attempts_all.columns[j][0:2]
    p_team = shot_attempts_all[p_id+'_team_id'].iloc[i]
    dist = shot_attempts_all.iloc[i,j]
    player = shot_attempts_all[p_id+'_id'].iloc[i]
    x_loc = shot_attempts_all[p_id+'_x_loc'].iloc[i]
    y_loc = shot_attempts_all[p_id+'_y_loc'].iloc[i]
    if (p_team != shooter_team) and (dist < min_dist):
      min_dist = dist
      p_id_closest = player
      x_loc_closest = x_loc
      y_loc_closest = y_loc
  closest_dist.append(min_dist)
  closest_player.append(p_id_closest)
  closest_x.append(x_loc_closest)
  closest_y.append(y_loc_closest)

shooter_x = []
shooter_y = []
for i in range(len(shot_attempts_all)):
  shooter = shot_attempts_all['shooter_id'].iloc[i]
  for j in range(1,11,1):
    p_id = shot_attempts_all['p'+str(j)+'_id'].iloc[i]
    if (shooter == p_id):
      x_loc = shot_attempts_all['p'+str(j)+'_x_loc'].iloc[i]
      y_loc = shot_attempts_all['p'+str(j)+'_y_loc'].iloc[i]
      break
  shooter_x.append(x_loc)
  shooter_y.append(y_loc)

shot_attempts_all['p_id_nearest_def'] = closest_player
shot_attempts_all['nearest_def_dist'] = closest_dist
shot_attempts_all['nearest_def_x_loc'] = closest_x
shot_attempts_all['nearest_def_y_loc'] = closest_y
shot_attempts_all['shooter_x_loc'] = shooter_x
shot_attempts_all['shooter_y_loc'] = shooter_y
shot_attempts_all.head()

shot_attempts_all.to_csv(cur_path+'shot_attempts_all_nearest_def.csv', index=False)
