==> 20-feature_engineering.Rmd <==
---
title: "feature engineering"
output: html_document
---

#load libraries
```{r load libraries, message=FALSE, warning=FALSE}

library(tidyverse)
library(janitor)
library(aspace)

options(scipen = 999)

`%not_in%` = Negate(`%in%`)

```

#load data
```{r load data, message=FALSE}
shot_attempts_all_nearest_def <- read_csv("data/shot_attempts_all_nearest_def.csv")
stats14_15 <- read_csv("data/stats14_15.csv")
stats15_16 <- read_csv("data/stats15_16.csv")
```


#create functions
```{r functions}
def_angle <- function(x, y, x2, y2){
  nearest_def_dist <- sqrt(((x-x2)^2) + ((y-y2)^2))
  shot_dist_from_L_hoop <- sqrt(((x-5.35)^2) + ((y-25)^2))
  shot_dist_from_R_hoop <- sqrt(((x-88.65)^2) + ((y-25)^2))
  def_dist_from_L_hoop <- sqrt(((x2-5.35)^2) + ((y2-25)^2))
  def_dist_from_R_hoop <- sqrt(((x2-88.65)^2) + ((y2-25)^2))
  shooter_to_hoop <- ifelse(shot_dist_from_L_hoop < shot_dist_from_R_hoop, shot_dist_from_L_hoop, shot_dist_from_R_hoop)
  def_hoop <- ifelse(def_dist_from_L_hoop < def_dist_from_R_hoop, def_dist_from_L_hoop, def_dist_from_R_hoop)
  def_angle <- acos_d(((nearest_def_dist^2)+(shooter_to_hoop^2)-(def_hoop^2))/(2*nearest_def_dist*shooter_to_hoop))
  return(def_angle)
  }

shooter_to_hoop <- function(x, y){
  shot_dist_from_L_hoop <- sqrt(((x-5.35)^2) + ((y-25)^2))
  shot_dist_from_R_hoop <- sqrt(((x-88.65)^2) + ((y-25)^2))
  shooter_to_hoop <- ifelse(shot_dist_from_L_hoop < shot_dist_from_R_hoop, shot_dist_from_L_hoop, shot_dist_from_R_hoop)
  return(shooter_to_hoop)
  }
```


```{r more cleaning}

shot_attempts_all_nearest_def <- shot_attempts_all_nearest_def %>%
  mutate(shooter_to_hoop = shooter_to_hoop(shooter_x_loc, shooter_y_loc),
         wrong_distance_flag = case_when(shot_zone_basic == "Above the Break 3" & shooter_to_hoop < 22 ~ 1,
                                         shot_zone_basic == "In The Paint (Non-RA)" & shooter_to_hoop < 2 ~ 1,
                                         shot_zone_basic == "In The Paint (Non-RA)" & shooter_to_hoop > 17 ~ 1,
                                         shot_zone_basic == "Left Corner 3" & shooter_to_hoop < 20.9 ~ 1,
                                         shot_zone_basic == "Right Corner 3" & shooter_to_hoop < 20.9 ~ 1,
                                         shot_zone_basic == "Mid-Range" & shooter_to_hoop < 2 ~ 1,
                                         shot_zone_basic == "Mid-Range" & shooter_to_hoop > 25.5 ~ 1,
                                         shot_zone_basic == "Restricted Area" & shooter_to_hoop > 6 ~ 1,
                                         TRUE ~ 0)
         ) %>% 
  filter(wrong_distance_flag == 0)

```


```{r create features}
shot_attempts_all_nearest_def <- shot_attempts_all_nearest_def %>% 
  mutate(points = case_when(shot_type == "3PT Field Goal" & event_type == "Made Shot" ~ 3,
                            shot_type == "2PT Field Goal" & event_type == "Made Shot" ~ 2,
                            event_type == "Missed Shot" ~ 0
                            ),
         points_factor = as.factor(points),
         shot_type_num = ifelse(shot_type == "2PT Field Goal", 2, 3),
         shooter_id = as.character(shooter_id),
         made_missed = as.factor(ifelse(points == 2 | points == 3, 1, 0)),
         nearest_def_dist = sqrt(((shooter_x_loc-nearest_def_x_loc)^2) + ((shooter_y_loc-nearest_def_y_loc)^2)),
         def_angle = def_angle(shooter_x_loc, shooter_y_loc, nearest_def_x_loc, nearest_def_y_loc),
         def_angle = ifelse(is.na(def_angle), 91, def_angle),
         rim_jump = case_when(grepl("Dunk", shot_attempts_all_nearest_def$action_type, ignore.case = TRUE) ~ "rim",
                              grepl("Lay", shot_attempts_all_nearest_def$action_type, ignore.case = TRUE) ~ "rim",
                              TRUE ~ "jump"
                              ),
         nearest_def_dist_round = case_when(nearest_def_dist <= .25 ~ 0,
                                            nearest_def_dist > .25 & nearest_def_dist <= .75 ~ .5,
                                            nearest_def_dist > .75 & nearest_def_dist <= 1.25 ~ 1,
                                            nearest_def_dist > 1.25 & nearest_def_dist <= 1.75 ~ 1.5,
                                            nearest_def_dist > 1.75 & nearest_def_dist <= 2.25 ~ 2,
                                            nearest_def_dist > 2.25 & nearest_def_dist <= 2.75 ~ 2.5,
                                            nearest_def_dist > 2.75 & nearest_def_dist <= 3.25 ~ 3,
                                            nearest_def_dist > 3.25 & nearest_def_dist <= 3.75 ~ 3.5,
                                            nearest_def_dist > 3.75 & nearest_def_dist <= 4.25 ~ 4,
                                            nearest_def_dist > 4.25 & nearest_def_dist <= 4.75 ~ 4.5,
                                            nearest_def_dist > 4.75 & nearest_def_dist <= 5.25 ~ 5,
                                            nearest_def_dist > 5.25 & nearest_def_dist <= 5.75 ~ 5.5,
                                            nearest_def_dist > 5.75 & nearest_def_dist <= 6.25 ~ 6,
                                            nearest_def_dist > 6.25 & nearest_def_dist <= 6.75 ~ 6.5,
                                            nearest_def_dist > 6.75 & nearest_def_dist <= 7.25 ~ 7,
                                            nearest_def_dist > 7.25 & nearest_def_dist <= 7.75 ~ 7.5,
                                            nearest_def_dist > 7.75 & nearest_def_dist <= 8.25 ~ 8,
                                            nearest_def_dist > 8.25 ~ 9,
                                            )
         ) %>% 
  group_by(shooter_id) %>% 
  mutate(n = n(), pps = sum(points)/n) %>% 
  ungroup()


```


```{r}
stats14_15 <- stats14_15 %>% 
  clean_names() %>% 
  select(-rk, fgp = fg_percent, tpm = x3p, tpa = x3pa, tpp = x3p_percent, tpfgm = x2p, tpfga = x2pa, tpfgp = x2p_percent, efg = e_fg_percent)


stats15_16 <- stats15_16 %>% 
  clean_names() %>% 
  select(-rk, fgp = fg_percent, tpm = x3p, tpa = x3pa, tpp = x3p_percent, tpfgm = x2p, tpfga = x2pa, tpfgp = x2p_percent, efg = e_fg_percent)

```


```{r}
shot_attempts_all_stats <- shot_attempts_all_nearest_def %>% 
  left_join(stats14_15, by = c('player_name' = 'player')) %>% 
  filter(!is.na(efg))

names_for_15_16 <- shot_attempts_all_nearest_def %>% 
  left_join(stats14_15, by = c('player_name' = 'player')) %>% 
  filter(is.na(efg)) %>% 
  select(player_name)

shot_attempts_all_stats_15_16 <- shot_attempts_all_nearest_def %>% 
  filter(player_name %in% names_for_15_16$player_name) %>% 
  left_join(stats15_16, by = c('player_name' = 'player'))

shot_attempts_all_stats <- bind_rows(shot_attempts_all_stats, shot_attempts_all_stats_15_16)


```






```{r}
less_than_two <- shot_attempts_all_stats %>% 
  mutate(ball_to_shooter = sqrt(((shooter_x_loc-ball_x_loc)^2) + ((shooter_y_loc-ball_y_loc)^2))) %>% 
  arrange(-ball_to_shooter) %>% 
  select(ball_to_shooter, shooter_id, shooter_x_loc, shooter_y_loc, shot_team_id, ball_x_loc, ball_y_loc, everything()) %>% 
  filter(ball_to_shooter < 2)

```


```{r}
write.csv(less_than_two, 'data/less_than_two.csv', row.names = FALSE)
```
