==> 30-modeling.Rmd <==
---
title: "modeling"
output: html_document
---

```{r load libraries, message=FALSE, warning=FALSE}
library(tidyverse)
library(janitor)
library(moderndive)
library(rsample)
library(randomForest)
library(caTools)
library(caret)
library(gbm)
library(Metrics)
library(sjlabelled)
library(ROCR)
#install.packages("rpart")
library(e1071)
library(rpart)

options(scipen = 999)

`%not_in%` = Negate(`%in%`)
```

```{r load data, message=FALSE}
less_than_two <- read_csv("data/less_than_two.csv") %>% 
  filter(shot_zone_basic != "Backcourt")
```


```{r}

shots_split <- initial_split(less_than_two)
shots_train <- training(shots_split)
shots_test <- testing(shots_split)

```


# baseline linear regression model
```{r}

baseline_model <- lm(points ~ shot_type_num * efg * fga + shot_type_num + shot_zone_basic + shooter_to_hoop + rim_jump + nearest_def_dist_round * def_angle, data = shots_train)

get_regression_summaries(baseline_model)
summary(baseline_model)
```

```{r}
y <- predict(baseline_model, shots_test)

rmse <- sqrt(sum((y - shots_test$points)^2)/length(shots_test$points))
rmse

#hist(y)

```



#random forest
```{r}
features_only <- less_than_two %>% 
  select(points, shot_type_num, efg, fga, shot_type_num, shot_zone_basic, shooter_to_hoop, rim_jump, nearest_def_dist_round, def_angle)

sapply(features_only, class)

features_only <- features_only %>% 
  mutate(shot_zone_basic = as.factor(shot_zone_basic),
         rim_jump = as.factor(rim_jump)
         )

features_only %>% 
  filter_all(any_vars(is.na(.)))


features_split <- initial_split(features_only)
features_train <- training(features_split)
features_test <- testing(features_split)

```


```{r}
#regression

rf_model <- randomForest(points ~ ., data = features_train, do.trace = TRUE)

y_rf <- predict(rf_model, features_test)

rmse_rf <- sqrt(sum((y_rf - features_test$points)^2)/length(features_test$points))
rmse_rf
```

```{r}
#fclassification

factor_points <- features_only %>% 
  mutate(points = as.factor(points))

factor_points_split <- initial_split(factor_points)
factor_points_train <- training(factor_points_split)
factor_points_test <- testing(factor_points_split)

factor_points_model <- randomForest(points ~ ., data = factor_points_train, do.trace = TRUE)

y_factor <- predict(factor_points_model, features_test)

confusionMatrix(y_factor, factor_points_test$points)
```




#stratisfy sample
```{r}
set.seed(123456)
shots_split_strat <- initial_split(less_than_two, prop = 3/5)
shots_train_strat <- training(shots_split_strat)
shots_test_strat <- testing(shots_split_strat)

#nrow(shots_train_strat)/nrow(less_than_two)
#nrow(shots_test_strat)/nrow(less_than_two)
```


```{r oversample}
# n rows of missed 2PT shots. Make this 1/4 of modeling df
missed_twos <- shots_train_strat %>% 
  filter(shot_type == "2PT Field Goal", made_missed == 0)

#made 2PT
made_twos <- shots_train_strat %>% 
  filter(shot_type == "2PT Field Goal", made_missed == 1)

made_twos_samp <- made_twos[sample(nrow(made_twos), nrow(missed_twos), replace = TRUE),]

#missed 3PT
missed_threes <- shots_train_strat %>% 
  filter(shot_type == "3PT Field Goal", made_missed == 0)

missed_threes_samp <- missed_threes[sample(nrow(missed_threes), nrow(missed_twos), replace = TRUE),]

#made 3PT
made_threes <- shots_train_strat %>% 
  filter(shot_type == "3PT Field Goal", made_missed == 1)

made_threes_samp <- made_threes[sample(nrow(made_threes), nrow(missed_twos), replace = TRUE),]

strat_sample <- bind_rows(missed_twos, made_twos_samp, made_threes_samp, missed_threes_samp)

train_strat_features <- strat_sample %>% 
  select(points, shot_type_num, efg, fga, shot_type_num, shot_zone_basic, shooter_to_hoop, rim_jump, nearest_def_dist_round, def_angle) %>% 
  mutate(points = as.factor(points),
         shot_zone_basic = as.factor(shot_zone_basic),
         rim_jump = as.factor(rim_jump)
  )

test_strat_features <- shots_test_strat %>% 
  select(points, shot_type_num, efg, fga, shot_type_num, shot_zone_basic, shooter_to_hoop, rim_jump, nearest_def_dist_round, def_angle) %>% 
  mutate(points = as.factor(points),
         shot_zone_basic = as.factor(shot_zone_basic),
         rim_jump = as.factor(rim_jump)
  )
```


```{r rf_strat}
strat_model <- randomForest(points ~ ., data = train_strat_features, do.trace = TRUE)
importance(strat_model)
y_strat <- predict(strat_model, test_strat_features)

confusionMatrix(y_strat, test_strat_features$points)
```


```{r cv}

#tried cross validating but results were significantly worse than several observed iterations of the same model and was unclear on interpretation, so cv not used

# numFolds <- trainControl(method = "cv", number = 10)
# cpGrid <- expand.grid(.cp = seq(0.01, 0.5, 0.01))
# train(points ~ ., data = train_strat_features, method = "rpart", trControl = numFolds, tuneGrid = cpGrid)
# 
# 
# rf_strat_cv <- rpart(points ~ ., data = train_strat_features, method = "class", cp = 0.01)
# rf_strat_cv_pred <- predict(rf_strat_cv, newdata = test_strat_features, type = "class")
# 
# confusionMatrix(rf_strat_cv_pred, test_strat_features$points)

```


```{r player level rf}
rf_labels_df <- as.data.frame(y_strat) %>% 
  mutate(rf_pred = as_numeric(y_strat)) %>% 
  select(rf_pred)

rf_pred_test <- bind_cols(rf_labels_df, shots_test_strat)

rf_pred_test_grouped <- rf_pred_test %>% 
  group_by(player_name) %>% 
  summarise(n = n(), pps = sum(points)/n, pps_pred = sum(rf_pred)/n, sample_efg = pps/2, pred_efg = pps_pred/2) %>% 
  arrange(-n)

rf_player_rmse <- rmse(actual = rf_pred_test_grouped$pps,
                  predicted = rf_pred_test_grouped$pps_pred)

rf_player_rmse

#omg it worked


write.csv(rf_pred_test, 'rf_pred_test.csv', row.names = FALSE) 
```








#gbm
```{r}
strat_regression_train <- strat_sample %>% 
  select(points, shot_type_num, efg, fga, shot_type_num, shot_zone_basic, shooter_to_hoop, rim_jump, nearest_def_dist_round, def_angle) %>% 
  mutate(shot_zone_basic = as.factor(shot_zone_basic),
         rim_jump = as.factor(rim_jump)
  )


strat_regression_test <- shots_test_strat %>% 
  select(points, shot_type_num, efg, fga, shot_type_num, shot_zone_basic, shooter_to_hoop, rim_jump, nearest_def_dist_round, def_angle) %>% 
  mutate(shot_zone_basic = as.factor(shot_zone_basic),
         rim_jump = as.factor(rim_jump)
  )
```

```{r}
strat_gbm <- gbm(points ~.,
             data = strat_regression_train,
             verbose = TRUE,
          shrinkage = 0.001, 
             # a shrinkage parameter applied to each tree in the expansion. 
             # Also known as the learning rate or step-size reduction; 0.001 
             # to 0.1 usually work, but a smaller learning rate typically 
             # requires more trees.
          interaction.depth = 2, 
             # Integer specifying the maximum depth of each tree (i.e., the 
             # highest level of variable interactions allowed). A value of 1 
             # implies an additive model, a value of 2 implies a model with up
             #  to 2-way interactions
          #n.minobsinnode = 5,
             # Integer specifying the minimum number of observations in the 
             # terminal nodes of the trees. Note that this is the actual number 
             # of observations, not the total weight.
          n.trees = 5000, 
             # Integer specifying the total number of trees to fit. This is 
             # equivalent to the number of iterations and the number of basis 
             # functions in the additive expansion.
             cv.folds = 10
             # Number of cross-validation folds to perform. If cv.folds>1 then
             # gbm, in addition to the usual fit, will perform a 
             # cross-validation, calculate an estimate of generalization error
             #  returned in cv.error
             )
```

```{r}
perf_strat_gbm = gbm.perf(strat_gbm, method = "cv")
```

```{r}
gbm_predictions <- predict(strat_gbm,
                           newdata = strat_regression_test,
                           n.trees = perf_strat_gbm)

rmse_gbm <- rmse(actual = strat_regression_test$points,
                  predicted = gbm_predictions)
rmse_gbm
```

```{r}
gbm_labels_df <- as.data.frame(gbm_predictions) %>% 
  mutate(gbm_pred = as_numeric(gbm_predictions)) %>% 
  select(gbm_pred)

gbm_pred_test <- bind_cols(gbm_labels_df, shots_test_strat)

gbm_pred_test_grouped <- gbm_pred_test %>% 
  group_by(player_name) %>% 
  summarise(n = n(), pps = sum(points)/n, pps_pred = sum(gbm_pred)/n, sample_efg = pps/2, pred_efg = pps_pred/2) %>% 
  arrange(-n)

gbm_player_rmse <- rmse(actual = gbm_pred_test_grouped$pps,
                  predicted = gbm_pred_test_grouped$pps_pred)

gbm_player_rmse

write.csv(gbm_pred_test, 'gbm_pred_test.csv', row.names = FALSE)
```


```{r}
summary(strat_gbm)
```







```{r}
gbm_classifier = gbm(points ~.,
              data = train_strat_features,
              distribution = "multinomial",
              cv.folds = 10,
              shrinkage = .01,
              #n.minobsinnode = 10,
              n.trees = 1000)

gbm_class_predictions <- predict(gbm_classifier,
                           newdata = test_strat_features)

summary(gbm_classifier)

labels = colnames(gbm_class_predictions)[apply(gbm_class_predictions, 1, which.max)]
result = data.frame(test_strat_features$points, labels)

confusionMatrix(test_strat_features$points, as.factor(labels))

```


```{r}
rf_labels_df <- as.data.frame(y_strat) %>% 
  mutate(rf_pred = as_numeric(y_strat)) %>% 
  select(rf_pred)
  
rf_labels_df <- bind_cols(test_strat_features, rf_labels_df)

grouped_preds <- rf_labels_df %>% 
  mutate(points = as_numeric(points),
         shooter_to_hoop = round(shooter_to_hoop),
         shooter_to_hoop = ifelse(shooter_to_hoop > 30, 31, shooter_to_hoop),
         def_angle_bucket = case_when(def_angle <= 22.5 ~ "front",
                                      def_angle > 22.5 & def_angle <= 45 ~ "side",
                                      def_angle > 45 ~ "oop"
                                      ),
         nearest_def_dist_round_2 = round(nearest_def_dist_round),
         nearest_def_dist_round_2 = case_when(shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 22 ~ 7,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 23 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 23 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 23 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 24 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 24 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 24 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 25 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 25 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 25 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 26 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 26 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 26 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 27 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 27 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 27 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 28 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 28 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 28 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 29 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 29 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 29 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 30 & nearest_def_dist_round_2 < 3 ~ 1,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 30 & nearest_def_dist_round_2 >= 3 & nearest_def_dist_round_2 < 6 ~ 4,
                                              shot_zone_basic == "Above the Break 3" & shooter_to_hoop == 30 & nearest_def_dist_round_2 >= 6 ~ 9,
                                              TRUE ~ nearest_def_dist_round_2
                                              )
         ) %>% 
  group_by(shot_zone_basic, shooter_to_hoop, nearest_def_dist_round_2) %>% 
  summarise(n = n(), points_tot = sum(points), pred_tot = sum(rf_pred), pps = points_tot/n, pred_pps = pred_tot/n)

tabyl(grouped_preds$n)

grouped_preds_small <- grouped_preds %>% 
  filter(n < 10) %>% 
  group_by(shot_zone_basic, nearest_def_dist_round_2) %>% 
  summarise(n = sum(n), points_tot = sum(points_tot), pred_tot = sum(pred_tot), pps = points_tot/n, pred_pps = pred_tot/n)


grouped_preds_very_small <- grouped_preds_small %>% 
  filter(n < 10) %>% 
  group_by(shot_zone_basic) %>% 
  summarise(n = sum(n), points_tot = sum(points_tot), pred_tot = sum(pred_tot), pps = points_tot/n, pred_pps = pred_tot/n)


grouped_preds_big <- grouped_preds %>% 
  filter(n >= 10)

grouped_preds_all <- bind_rows(grouped_preds_big, grouped_preds_small, grouped_preds_very_small)



rmse_grouped <- rmse(actual = grouped_preds_all$pps,
                  predicted = grouped_preds_all$pred_pps)
print(rmse_gbm)
  
```









#lm with stratified sample
```{r}
strat_lm <- lm(points ~ shot_type_num * efg * fga + shot_type_num + shot_zone_basic + shooter_to_hoop + rim_jump + nearest_def_dist_round * def_angle, data = shots_train_strat)

get_regression_summaries(strat_lm)
summary(start_lm)
```

```{r}
strat_lm_y <- predict(strat_lm, shots_test_strat)

rmse_strat_lm <- sqrt(sum((strat_lm_y - shots_test_strat$points)^2)/length(shots_test_strat$points))
rmse_strat_lm

#hist(y)

```

```{r}
log_model_strat <- glm(made_missed ~ efg * fga * shooter_to_hoop + shot_zone_basic * shooter_to_hoop + shooter_to_hoop * rim_jump + efg * nearest_def_dist * def_angle + shooter_to_hoop * nearest_def_dist * def_angle, family = binomial (link = 'logit'), data = strat_sample)

summary(log_model_strat)


log_pred_strat <- predict(log_model_strat, newdata = shots_test_strat, type='response')
pred_results_strat <- ifelse(log_pred_strat >= 0.5, 1, 0)
misclasification_error_strat <- mean(pred_results_strat != shots_test_strat$made_missed)
print(paste("Accuracy:", round(1-misclasification_error_strat, 4)))
```



```{r logistic reg}
log_model <- glm(made_missed ~ efg * fga * shooter_to_hoop + shot_zone_basic * shooter_to_hoop + shooter_to_hoop * rim_jump + efg * nearest_def_dist * def_angle + shooter_to_hoop * nearest_def_dist * def_angle, family = binomial (link = 'logit'), data = shots_train)

summary(log_model)


log_pred <- predict(log_model, newdata = shots_test, type='response')
pred_results <- ifelse(log_pred >= 0.5, 1, 0)
misclasification_error <- mean(pred_results != shots_test$made_missed)
print(paste("Accuracy:", round(1-misclasification_error, 4)))

#confusionMatrix(pred_results, shots_test$made_missed)

```

```{r}

pr <- prediction(log_pred, shots_test$made_missed)
prf <- performance(pr, measure = "tpr", x.measure = "fpr")
plot(prf)
auc <- performance(pr, measure = "auc")
auc <- auc@y.values[[1]]
auc
```



```{r}
log_labels_df <- as.data.frame(pred_results)

log_labels_df <- bind_cols(shots_test, log_labels_df)

log_labels_df <- log_labels_df %>% 
  mutate(pred_points = pred_results * shot_type_num)

log_labels_df %>% 
  group_by(player_name) %>% 
  summarise(n = n(), pps = sum(points)/n, pps_pred = sum(pred_points)/n, sample_efg = pps/2, pred_efg = pps_pred/2) %>% 
  arrange(-n)


log_labels_df %>% 
  filter(player_name == "Stephen Curry") %>% 
  group_by(shot_type, made_missed) %>% 
  summarise(n = n())
```

```{r}
(25+(1.5*39))/110
```
