==> 40-analysis.Rmd <==
---
title: "analysis"
output: html_document
---

```{r message=FALSE}
library(tidyverse)
library(janitor)
library(sm)
library(kableExtra)


rf_pred_test <- read_csv("rf_pred_test.csv")
stats16_17 <- read_csv("data/stats16_17.csv") %>% 
  clean_names()
```


```{r}
#group to player level and create aggregate eFG% features
gbm_pred_test_grouped_2 <- gbm_pred_test_2 %>% 
  group_by(player_name) %>% 
  summarise(n = n(), pps = sum(points)/n, pps_pred = sum(gbm_pred)/n, sample_efg = pps/2, pred_efg = pps_pred/2) %>% 
  arrange(-n)
```

```{r}
#histogram showing distribution of predictions

png(file = "pred_distribution.png")
hist(gbm_predictions_2,
     main = "Distribution of Predictions",
     xlab = "Predicted Shot Values"
     )
dev.off()


```


```{r}
#double density plot

ggplot() +
  geom_density(aes(x = sample_efg), colour = "red", data = filter_50) + 
  geom_density(aes(x = pred_efg), colour = "blue", data = filter_50) +
  theme_classic() +
  theme(axis.ticks.y=element_blank(),
        axis.text.y=element_blank(),
        ) +
  labs(y = "Density", x = "Red: Observed eFG% - Blue: Predicted eFG%") +
  ggtitle("Observed vs. Predicted eFG%") +
  theme(plot.title = element_text(hjust = 0.5))
  
ggsave("observed_vs_pred.png")

```


```{r}
#cleaning summary output for plot

gbm_summary <- gbm_summary %>% 
  mutate(Feature = case_when(var == "shot_zone_basic" ~ "Shot Location",
                         var == "nearest_def_dist_round" ~ "Distance from Nearest Defender",
                         var == "shooter_to_hoop" ~ "Shot Distance",
                         var == "efg" ~ "Shooter eFG% 2014-15",
                         var == "def_angle" ~ "Angle of Defender to Shooter",
                         var == "rim_jump" ~ "Jump Shot or Play at Rim",
                         var == "fga" ~ "Shooter Total FGA 2014-15",
                         var == "shot_type_num" ~ "2pt or 3pt FGA",
                         ),
         'Relative Influence' = round(rel.inf, 2)
  ) %>% 
  select(Feature, 'Relative Influence')

```


```{r}
#plotting gbm features

ggplot(gbm_summary, aes(x = reorder(Feature, `Relative Influence`), y = `Relative Influence`)) +
  geom_bar(stat = "identity", fill = 'lightseagreen', colour = "black") +
  xlab("Feature") +
  coord_flip() +
  theme_classic()
  

ggsave('features.png', height = 7, width = 10.5)
```


```{r}

sd(gbm_pred_test_grouped_2$sample_efg)
```


```{r}
#plot volume against difference in prediction
ggplot(gbm_pred_test_grouped_2, aes(x = gbm_pred_test_grouped_2$n, y = gbm_pred_test_grouped_2$difference)) +
  geom_point(size = 3, alpha = .75, col = "lightseagreen") +
  geom_smooth(col = "black", method = lm, se = FALSE) +
  theme_classic() +
  labs(x = "Shot Volume", y = "Difference in Predicted vs. Observed eFG%") +
  ggtitle("Difference in Observed vs. Predicted eFG% by Shot Volume") +
  theme(plot.title = element_text(hjust = 0.5))

ggsave('volume.png', height = 7, width = 10.5)
  
```


```{r}
filter_100 <- gbm_pred_test_grouped_2 %>% 
  mutate(difference = pred_efg - sample_efg) %>% 
  filter(n >= 100) %>% 
  arrange(difference)
```

```{r}
#create df showing predicted and observed increases or decreases
filter_100 %>% 
  left_join(stats16_17, by = c('player_name' = 'player')) %>% 
  select(player_name, n, sample_efg, pred_efg, difference, x1617_efg = e_fg_percent) %>% 
  filter(!is.na(x1617_efg)) %>% 
  mutate(increase_pred = ifelse(difference < 0, "decrease",  "increase"), increase_1617 = ifelse(sample_efg < x1617_efg, "increase", "decrease"), sample_efg = round(sample_efg, 3), pred_efg = round(pred_efg, 3), difference = round(difference, 3)) %>% 
  rename(Player = player_name, FGA = n, 'Sample eFG%' = sample_efg, "Pred. eFG%" = pred_efg, Diff. = difference, '16-17 eFG%' = x1617_efg, Predicted = increase_pred, Observed = increase_1617) %>% 
  arrange(-Diff.) %>% 
  head(10) %>% 
  kable() %>% 
  kable_styling(position = "center", full_width = TRUE)
```
